[LEMS-2851/answerless-numeric-input] move utils out of perseus
