+++ packages/perseus/src/widgets/numeric-input/numeric-input.class.tsx
@@ -1,5 +1,8 @@
 import {KhanMath} from "@khanacademy/kmath";
-import {getNumericInputPublicWidgetOptions} from "@khanacademy/perseus-core";
+import {
+    deriveAnswerForms,
+    getNumericInputPublicWidgetOptions,
+} from "@khanacademy/perseus-core";
 import {linterContextDefault} from "@khanacademy/perseus-linter";
 import * as React from "react";
 
@@ -7,7 +10,6 @@ import {ApiOptions} from "../../perseus-api";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 
 import {NumericInputComponent} from "./numeric-input";
-import {deriveAnswerForms} from "./utils";
 
 import type InputWithExamples from "./input-with-examples";
 import type SimpleKeypadInput from "../../components/simple-keypad-input";
